% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_to_save_down_deposit.R
\name{m_to_save_down_deposit}
\alias{m_to_save_down_deposit}
\title{Calculate months to save down deposit}
\usage{
m_to_save_down_deposit(salary, house_cost, prop_saved, ret_rate)
}
\arguments{
\item{salary}{Annual salary, in dollars}

\item{house_cost}{The cost of the house to be purchased}

\item{prop_saved}{The proportion of monthly salary set aside for investment every month}

\item{ret_rate}{The annual return rate on investments}
}
\value{
Returns the numbers of months needed to save for a 25\% downpayment on a house
}
\description{
Calculates the number of months needed to save for a 25\% down-payment on a house, based on current
income, the cost of the house, the proportion of monthly income invested every month, and the annual
investment return rate. Assumes that the initial amount of savings is 0.
}
\examples{
m_to_save_down_deposit(salary=80000, house_cost=2000000, prop_saved=0.05, ret_rate=0.05)
m_to_save_down_deposit(salary=150000, house_cost=1500000, prop_saved=0.25, ret_rate=0.02)
}
